Remove lint command (#4881)
* remove lint and xsoar_linter

* add DockerImageFlagOption

* add excluded_files

* import enum

* add resources dir

* run ruff

* update changelog

* update changelog

* update requirements

* update pre commit resources

* ruff

* checkout remove-lint branch

* remove consts

* put back xsoar-linter

* put back xsoar-linter

* put back xsoar-linter

* put back consts

* fix test_xsoar_linter_errors

* pre commit

* remove xsoar_linter_integration_test.py
