+++ .changelog/4881.yml
@@ -1,0 +1,4 @@
+changes:
+- description: Removed ***lint*** command.
+  type: breaking
+pr_number: 4881
