+++ demisto_sdk/commands/pre_commit/hooks/docker.py
@@ -6,6 +6,7 @@ import time
 from collections import defaultdict
 from concurrent.futures import ThreadPoolExecutor
 from copy import deepcopy
+from enum import Enum
 from functools import lru_cache
 from pathlib import Path
 from typing import Dict, Iterable, List, Optional, Set, Tuple
@@ -38,8 +39,19 @@ from demisto_sdk.commands.common.tools import logger
 from demisto_sdk.commands.content_graph.objects.integration_script import (
     IntegrationScript,
 )
-from demisto_sdk.commands.lint.linter import DockerImageFlagOption
 from demisto_sdk.commands.pre_commit.hooks.hook import GeneratedHooks, Hook
+
+
+class DockerImageFlagOption(Enum):
+    FROM_YML = "from-yml"
+    NATIVE = "native:"
+    NATIVE_DEV = "native:dev"
+    NATIVE_GA = "native:ga"
+    NATIVE_MAINTENANCE = "native:maintenance"
+    ALL_IMAGES = "all"
+    NATIVE_TARGET = "native:target"
+    NATIVE_CANDIDATE = "native:candidate"
+
 
 NO_SPLIT = None
