+++ demisto_sdk/commands/pre_commit/resources/installation_scripts/powershell_image.sh
