+++ demisto_sdk/commands/pre_commit/resources/installation_scripts/python_image.sh
