+++ demisto_sdk/commands/pre_commit/resources/pylint_plugins/base_checker.py
