+++ demisto_sdk/commands/pre_commit/resources/pylint_plugins/certified_partner_level_checker.py
