+++ demisto_sdk/commands/pre_commit/resources/pylint_plugins/community_level_checker.py
