+++ demisto_sdk/commands/pre_commit/resources/pylint_plugins/partner_level_checker.py
