+++ demisto_sdk/commands/pre_commit/resources/pylint_plugins/xsoar_level_checker.py
