+++ demisto_sdk/commands/pre_commit/resources/python2_requirements/dev-requirements.txt
