+++ demisto_sdk/commands/pre_commit/resources/python3_requirements/dev-requirements.txt
